#5985: Splitting out trace code from dispatch/command_queue, it is dispatch agnostic
